chat scrollbar fixed to bottom, fixed UserStats bug
