+++ frontend/volume/src/components/chat/JoinPrivate.tsx
@@ -1,0 +1,9 @@
+import "../../styles/style.css";
+
+export default function PrivateChannel() {
+	return (
+		<div id="new_chat">
+			<h2>TO DO: list of public channels</h2>
+		</div>
+	);
+}
