+++ frontend/volume/src/components/common/UserStats.tsx
@@ -11,8 +11,25 @@ import { useOutgoingRequests } from "src/utils/useOutgoingRequests";
 
 export default function UserStats(modalProps: i.ModalProps & { selectedUser: any }) {
 	const { friends, loading, error } = useFriendsData(modalProps.userId);
-	if (loading) return <div>Loading friends</div>;
+	const [remove_friend, { loading: loadingRemove, error: errorRemove }] =
+		useMutation(REMOVE_FRIEND);
+	const [invite_friend, { loading: loadingRequest, error: errorRequest }] =
+		useMutation(INVITE_FRIEND);
+	const {
+		outgoingRequests,
+		loading: loadingOutgoing,
+		error: errorOutgoing,
+	} = useOutgoingRequests(modalProps.userId);
+
+	if (loading) return <> </>;
+	if (loadingOutgoing) return <div>Loading friends</div>;
+	if (loadingRemove) return <> </>;
+	if (loadingRequest) return <>send friend request</>;
+
 	if (error) return <div>Error friends</div>;
+	if (errorOutgoing) return <>error</>;
+	if (errorRemove) return <>error</>;
+	if (errorRequest) return <>error</>;
 
 	const renderUserActions = () => {
 		if (modalProps.selectedUser.id === modalProps.userId)
@@ -34,7 +51,13 @@ export default function UserStats(modalProps: i.ModalProps & { selectedUser: any
 				>
 					challenge
 				</a>
-				{renderFriendRequestActions(friends, modalProps)}
+				{renderFriendRequestActions(
+					friends,
+					modalProps,
+					outgoingRequests,
+					remove_friend,
+					invite_friend
+				)}
 				{renderBlockOrUnblock(friends, modalProps)}
 			</div>
 		);
@@ -67,21 +90,13 @@ const REMOVE_FRIEND = gql`
 	}
 `;
 
-function renderFriendRequestActions(friends: any, modalProps: any) {
-	const [remove_friend, { loading: loadingRemove, error: errorRemove }] =
-		useMutation(REMOVE_FRIEND);
-	const [invite_friend, { loading: loadingRequest, error: errorRequest }] =
-		useMutation(INVITE_FRIEND);
-	const { outgoingRequests, loading, error } = useOutgoingRequests(modalProps.userId);
-
-	if (loading) return <> </>;
-	if (loadingRemove) return <> </>;
-	if (loadingRequest) return <>send friend request</>;
-
-	if (error) return <>error</>;
-	if (errorRemove) return <>error</>;
-	if (errorRequest) return <>error</>;
-
+function renderFriendRequestActions(
+	friends: any,
+	modalProps: any,
+	outgoingRequests: any,
+	remove_friend: any,
+	invite_friend: any
+) {
 	if (outgoingRequests.find((request: any) => request.id === modalProps.selectedUser.id))
 		return <>friend request has been sent</>;
 	if (friends.find((friend: any) => friend.id === modalProps.selectedUser.id))
